open-testimony-app/bridge/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.3.6
minio==7.2.3
python-jose[cryptography]==3.3.0
Pillow>=10.0.0
opencv-python-headless>=4.6.0
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.24.0
sentence-transformers>=2.7.0
pywhispercpp>=1.3.0
numpy>=1.24.0
httpx>=0.25.0
